# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Demo Quizzes",
        "description": "Quizzes API",
        "contact": {
            "email": "huan.ld@neo-lab.vn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Quizzes API"
        }
    ],
    "paths": {
        "/api/admin/questions": {
            "post": {
                "tags": [
                    ""
                ],
                "summary": "Add Question Admin",
                "operationId": "storeQuestions",
                "parameters": [
                    {
                        "name": "topic_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "question_text",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code_snippet",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "answer_explanation",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "more_info_link",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "option[0]",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "option[1]",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "option[2]",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "option[3]",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "correct",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Token not provided"
                    },
                    "200": {
                        "description": "Created Question Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/topics": {
            "get": {
                "tags": [
                    ""
                ],
                "summary": "List Topics Admin",
                "operationId": "viewTopics",
                "responses": {
                    "401": {
                        "description": "Token not provided"
                    },
                    "200": {
                        "description": "List Topics Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    ""
                ],
                "summary": "Create Topics Admin",
                "operationId": "doTopic",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Token not provided"
                    },
                    "200": {
                        "description": "Create Topic Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/list-action-users": {
            "get": {
                "tags": [
                    ""
                ],
                "summary": "List Action of User",
                "operationId": "listAction",
                "responses": {
                    "401": {
                        "description": "Token not provided"
                    },
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}